Start separation of ehtim
